Make ChangeScreenSettings imply InterruptSession
As e.g. suspending the computer will interrupt video playing. There do
not seem to be any use cases that require keeping the screen active
right up to the moment that the system idles out and suspends itself.
Indeed, PolicyAgent users that ask for ChangeScreenSettings inhibit do
not expect the screen contents to disappear, no matter the cause.

REVIEW: 107699