Fix bug 289427 - plasma-desktop freezes when selecting fixed-size fonts in the clock widget
Bug 289427 is caused by a endless loop when searching for largest suitable font.
If the largest size of current system font is still smaller than the shown area,
the font.pointSize is increased but the resulting size QRect is the same.
So the do...while loop will loop forever and consume 100% cpu runtime.

Fix:
Check whether the resulting QRect is still increasing. If not, end the loop.

BUG: 289427
REVIEW: 108102
FIXED-IN: 4.10.0