NepomukRunner: Redirect the 'nepomuk:/res/uuid' urls before opening them
The nepomuksearch kioslave has a special syntax which results in the the
searched term being displayed in "human terms" without the need for
stating the url. The nepomuk kioslave does not have this feature. It,
however, uses the nepomuksearch kioslave internally.

Applications like Dolphin often do not stat a url before displaying it
in the title bar. This results in users seing an ugly "nepomuk - res -
uuid" in the title bar instead of the proper display name as set in the
UDSEntry.

In order to avoid this ugly title bar, we manually redirect all nepomuk
urls to nepomuksearch before they are opened by the runner.

It's a hack, but it works!

BUG: 287259
REVIEW: 108305