NepomukRunner: Do not show ontology results
The Nepomuk Query Engine should typically never return ontology results,
but unfortunately it does. We have tried different ways of filtering
them out - Earlier attempts included complex queries (nao:userVisible)
which results in the queries becoming very complex and virtuoso acting
up.

The current method to filter them is via graph groups. Unfortunately,
there is a bug in the current virtuoso which causes it to crash when
using graph groups. So, the policy for 4.10 is just to filter the
results in an adhoc manner in the client code.