Stop grace period on unlock
It was safe to leak the grace timer when it was in the lock process,
since it was killed on unlock. Now it is not.

A password should always be required if the screenlocker is explicitly
triggered; but a screenlocker invoked while the grace timer from a
previous invocation is still active can be dismissed without a password,
at least for a short while. This also meant that the screenlocker
keyboard shortcut would not work until the grace timer expired.

REVIEW: 108425
BUG: 313997
FIXED-IN: 4.10.1