Fix handling of encoded URIs in mpris2 engine
A couple of entries in the metadata property of the MPRIS2 interface are
sent as encoded URIs.  We decode these in the mpris2 engine, replacing
the QStrings with QUrls that can be used easily by the consumer of the
engine.

This fixes the display of album artwork in the Now Playing widget when
mpris:artUrl contains %-encoded non-ASCII characters.

BUG: 303088
FIXED-IN: 4.10