Also set the font's style name when batch-changing all fonts.
This fixes a regression created by Qt 4.8: calling KFontRequester::setFont()
ends up calling the nearestExistingFont() helper function, which in turn
calls QFontDatabase::styleString(). In Qt 4.8, QFontDatabase::styleString()
calls QFont::styleName() before falling back to the old code path.

In our case, we ended up inheriting the style name of the current "General"
font.

In practice, this means that if the current "General" font had a style such
as "Regular", it was not possible to apply another one such as "Light" or
even "Bold" to all fonts because the other properties we set were
disregarded in favour of QFont::styleName().

Changing each font individually worked fine because it uses a different code
path.