mpris2: Find out the media player icon if available
We can't really expect applets to do desktop file lookups, so if there
is a DesktopEntry property, we use it to make a "Desktop Icon Name"
property.  This should only be done when the media player first appears,
so the overhead of reading the desktop file should not be noticable.