update geometry of the lock window when screen changes
this is the window that lays *under* the acdtual lock UI, and which
prevents visibility of the underlying desktop (and lets the window
manager know what's up) thereby allowing us to care a lot less about
whether the greeter crashes while improving security with compositing on.

previously it was not changing geom when the screen would .. which was
not noticed due to a combination of issues; but with kscreen being used,
its magic revealed the problem.

based on a patch by Thomas Lübking

REVIEW:108643
REVIEW:108643
BUG:108643