Add Nepomuk-Controller QML applet and DataEngine
The Nepomuk DataEngine allows to start/stop/suspend/resume/configure
the nepomuk indexer services and retuen some basic information about
them.

The QMl applet is ment to replace the current systray controller to
allow the user to see what nepomuk is doing in general. This helps
each user to understand why nepomuk is working and also to suspend
all tasks if the user needs his cpu resources at the moment.