pop in an assert on the pointer
it was checked once after retrieval from the smart pointer, but it was never
checked later in the method.

the smart pointer is nearly useless for runtime safety as used here,
but at least it will give us nice backtraces with 0 pointers if the code
ever changes to void the current assumptions in the code that the model
exists as long as the config UI does.