Make the startup of the greeter async
This way we don't block the event loop waiting for the process to write
on its standard output. Problem now is of course in the case of suspend
the locker will be brought up only on resume, next commit will solve
that.

Also we don't really need KProcess here so port to QProcess in the
process (aha).