Plasma Wallpaper BackgroundListModel: Avoid using KFileMetaInfo
1. The Xesam width and height properties are never returned by strigi
and haven't been returned for a very very long time. It currently
returns the nfo:width and nfo:height property instead.

2. KFileMetadataWidget is blocking and blocks while it is calculating
the height and width. It is better to open the file and calculate it in
another thread as done by the ImageSizeFinder

3. QImage in more reliable and actively maintained in contrast to
strigi.

REVIEW: 110467
(cherry picked from commit 8a6f783dfd6f021173f5d3b47772b0fc320eba31)

This change was somehow lost in the merges

Conflicts:
	plasma/generic/wallpapers/image/backgroundlistmodel.cpp