Modify Tasks applet to display in which activities a window is on
Shows information about the activities on which a window is available on the
tooltip of task items.

Depending whether the applet is configured to show tasks which are not
available in the current activity or not, the information displayed will be
different.
If the applet is configured to show only tasks in current activity, only the
tasks which are also available in other activities will have this information
in the tooltip.

Two new methods were added to TaskItem, both of them returning QStringList
 - activities: Returns the id of the activities on which the window is
currently available.
 - activityNames: Returns the names of those activities.

BUG: 307163
FEATURE: 110413