KDisplayManager: Add a basic session creating test
Basically just a proof of concept.

It tests if it's possible to create a new session (to create a new greeter instance on another VT) and creates one

To do: cleanup after the action, switching back to the initial session. KWorkspace/KDisplayManager itself doesn't provide facilities to do that or more precisely, it's not always possible to do that using the switchVT method as it's more sane to allow session switching only to the currently active one.