Show new applications in a seperate submenu named "Recently Installed" (for 3 days).
Patch by Wolfgang Bauer.

This patch makes kickoff remember all the .desktop files it sees (in kickoffrc).
The seen entries are saved as "id=date" pairs in the "Seen Applications" group.
Entries older than 3 days get their date set to empty, which means "don't show this entry anymore".
Also if there have no entries been saved yet, the date of all found entries is
set to empty as well. This prevents that all menu entries are shown in the
"Recently Installed" submenu on first start.

The "Recently Installed" submenu can be toggled on and off in the plasmoid's settings.