Wayland implementation of Screens interface
Uses the information provided by the Wayland outputs to setup the data
about the available screens.

In order to properly work together with the X Server the implementation
syncs the data to the X system using XRandR. If XRandR is not available
this is most likely going to not work correctly.