Hide unimplemented manual grouping option.
This is a feature regression vs. 4.10, but there's no time
left to implement it now. Fortunately, this had almost no
users anyway since discovering how to work it generally
required a Google search, and even then it was horribly
unfriendly to actually use - we can very likely do much
better next time we take a stab at the idea.